# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: company
    description: Contracting company system operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /company/overview:
    get:
      operationId: getCompanyOverview
      tags: [company]
      summary: Company overview
      responses:
        "200":
          description: Company operational overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyOverview"
  /company/employees:
    get:
      operationId: listEmployees
      tags: [company]
      summary: List employees
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [company]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /company/projects:
    get:
      operationId: listProjects
      tags: [company]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [company]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /company/contracts:
    get:
      operationId: listContracts
      tags: [company]
      summary: List contracts
      responses:
        "200":
          description: Contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
  /company/tasks:
    get:
      operationId: listOperationalTasks
      tags: [company]
      summary: List operational tasks
      responses:
        "200":
          description: Operational tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationalTask"
    post:
      operationId: createOperationalTask
      tags: [company]
      summary: Create operational task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperationalTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationalTask"
  /company/tasks/{id}/status:
    patch:
      operationId: updateOperationalTaskStatus
      tags: [company]
      summary: Update operational task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationalTask"
  /company/approvals:
    get:
      operationId: listApprovals
      tags: [company]
      summary: List approvals
      responses:
        "200":
          description: Approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
  /company/approvals/{id}/status:
    patch:
      operationId: updateApprovalStatus
      tags: [company]
      summary: Update approval status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusInput"
      responses:
        "200":
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /company/activity:
    get:
      operationId: listActivity
      tags: [company]
      summary: List recent activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /company/operations:
    get:
      operationId: listOperationalRecords
      tags: [company]
      summary: List operational records across company modules
      responses:
        "200":
          description: Operational records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationalRecord"
    post:
      operationId: createOperationalRecord
      tags: [company]
      summary: Create operational record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperationalRecordInput"
      responses:
        "201":
          description: Created operational record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationalRecord"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Metric:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
        change:
          type: string
      required: [label, value, change]
    CompanyOverview:
      type: object
      properties:
        activeProjects:
          type: integer
        pendingApprovals:
          type: integer
        monthlyCompletion:
          type: integer
        totalContractValue:
          type: number
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/Metric"
        departmentBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/Metric"
      required: [activeProjects, pendingApprovals, monthlyCompletion, totalContractValue, metrics, departmentBreakdown]
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        department:
          type: string
        status:
          type: string
        joinedAt:
          type: string
          format: date-time
      required: [id, name, email, phone, role, department, status, joinedAt]
    CreateEmployeeInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        department:
          type: string
      required: [name, email, phone, role, department]
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientName:
          type: string
        location:
          type: string
        stage:
          type: string
        status:
          type: string
        progress:
          type: integer
        budget:
          type: number
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
      required: [id, name, clientName, location, stage, status, progress, budget, startsAt, endsAt]
    CreateProjectInput:
      type: object
      properties:
        name:
          type: string
        clientName:
          type: string
        location:
          type: string
        stage:
          type: string
        budget:
          type: number
      required: [name, clientName, location, stage, budget]
    Contract:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        projectName:
          type: string
        clientName:
          type: string
        value:
          type: number
        paidAmount:
          type: number
        status:
          type: string
        signedAt:
          type: string
          format: date-time
      required: [id, code, projectName, clientName, value, paidAmount, status, signedAt]
    OperationalTask:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        projectName:
          type: string
        assignee:
          type: string
        priority:
          type: string
        status:
          type: string
        dueAt:
          type: string
          format: date-time
      required: [id, title, category, projectName, assignee, priority, status, dueAt]
    CreateOperationalTaskInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        projectName:
          type: string
        assignee:
          type: string
        priority:
          type: string
      required: [title, category, projectName, assignee, priority]
    UpdateStatusInput:
      type: object
      properties:
        status:
          type: string
      required: [status]
    Approval:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        requester:
          type: string
        projectName:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, type, requester, projectName, status, createdAt]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        actor:
          type: string
        module:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, actor, module, createdAt]
    OperationalRecord:
      type: object
      properties:
        id:
          type: integer
        module:
          type: string
        title:
          type: string
        reference:
          type: string
        projectName:
          type: string
        owner:
          type: string
        status:
          type: string
        priority:
          type: string
        quantity:
          type: integer
        amount:
          type: number
        dueAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, module, title, reference, projectName, owner, status, priority, quantity, amount, dueAt, createdAt]
    CreateOperationalRecordInput:
      type: object
      properties:
        module:
          type: string
        title:
          type: string
        reference:
          type: string
        projectName:
          type: string
        owner:
          type: string
        status:
          type: string
        priority:
          type: string
        quantity:
          type: integer
        amount:
          type: number
        dueAt:
          type: string
          format: date-time
      required: [module, title, reference, projectName, owner, status, priority, quantity, amount, dueAt]