# lib/api-zod/src/generated/types/createOperationalRecordInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOperationalRecordInput {
  module: string;
  title: string;
  reference: string;
  projectName: string;
  owner: string;
  status: string;
  priority: string;
  quantity: number;
  amount: number;
  dueAt: Date;
}